Refactor and fix some problems. Change structure
