+++ src/components/films/film-details/details-info/titles.js
@@ -1,7 +1,7 @@
 /**
- * Создание разметки с названиями фильма
+ * Создание разметки блока названий фильма
  * @param {Object} {переведенное и оригинальное названия}
- * @return {string} разметка названий
+ * @return {string} разметка блока
  */
 const createTitles = ({translate, original}) => {
   return (`
