+++ src/components/films/sorting.js
@@ -1,6 +1,6 @@
 /**
- * Создание шаблона панели сортировки
- * @return {string} сортировка
+ * Создание разметки блока сортировки фильмов
+ * @return {string} разметка блока
  */
 const createSorting = () => {
   return (`
