+++ src/components/stats.js
@@ -2,8 +2,8 @@ import {CountFilm} from "../consts";
 import {castNumberFormat} from "../utils";
 
 /**
- * Создание шаблона информации о статистике
- * @return {string} информация
+ * Создание разметки блока информации о количестве фильмов
+ * @return {string} разметка
  */
 const createStatistic = () => `<p>${castNumberFormat(CountFilm.ALL)} movies inside</p>`;
